# omakase/src/main/webapp/swagger/v1/omakase.json
{
  "basePath": "/omakase/api",
  "swagger": "2.0",
  "info": {
    "title": "Omakase",
    "version": "1.0"
  },
...
  "paths": {
...
    "/assets/{assetId}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "assets"
        ],
        "operationId": "getAssetById",
        "summary": "Retrieve an asset",
        "parameters": [
          {
            "in": "path",
            "description": "Asset Id",
            "name": "assetId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetAsset"
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "assets"
        ],
        "operationId": "deleteAsset",
        "summary": "Delete an asset",
        "parameters": [
          {
            "in": "path",
            "description": "Asset Id",
            "name": "assetId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseStatus"
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "assets"
        ],
        "operationId": "updateAsset",
        "summary": "Updates an asset",
        "description": "Only name and external id are updated all other fields are read only.",
        "parameters": [
          {
            "in": "path",
            "description": "Asset Id",
            "name": "assetId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GetAsset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseStatus"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "assets"
        ],
        "operationId": "partialUpdateAsset",
        "summary": "Partially updates an asset",
        "description": "Supports updating the asset name and adding and removing external ids using JSON Patch.",
        "consumes": [
          "application/json-patch+json"
        ],
        "parameters": [
          {
            "in": "path",
            "description": "Asset Id",
            "name": "assetId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JSON Patch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseStatus"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/assets/{assetId}/variants/{variantId}/files": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "assets"
        ],
        "operationId": "getFiles",
        "summary": "Retrieve the variant's files",
        "parameters": [
          {
            "in": "path",
            "description": "Asset Id",
            "name": "assetId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "description": "Variant Id",
            "name": "variantId",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "description": "The page of results to return",
            "name": "page",
            "required": false,
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "description": "The number of results per page",
            "name": "per_page",
            "required": false,
            "type": "integer",
            "format": "int32",
            "default": 10
          },
          {
            "in": "query",
            "description": "Determines the attribute used to sort the results",
            "name": "sort",
            "required": false,
            "type": "string",
            "default": "created"
          },
          {
            "in": "query",
            "description": "Determines the sort order of the results",
            "name": "order",
            "required": false,
            "type": "string",
            "default": "DESC"
          },
          {
            "in": "query",
            "description": "Limits the results to an aggregate, pagination is ignored.",
            "name": "only_count",
            "required": false,
            "type": "boolean",
            "default": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/FilesEnvelope"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/assets/{assetId}/variants/{variantId}/files/{fileId}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "assets"
        ],
        "operationId": "getFiles",
        "summary": "Retrieves a variant's file",
        "parameters": [
          {
            "in": "path",
            "description": "Asset Id",
            "name": "assetId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "description": "Variant Id",
            "name": "variantId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "description": "File Id",
            "name": "fileId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/File"
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ResponseStatus": {
      "required": [
        "status",
        "links"
      ],
      "properties": {
        "status": {
          "type": "string",
          "enum": [
            "OK",
            "ERROR"
          ]
        },
        "message": {
          "type": "string"
        },
        "exception": {
          "type": "string"
        },
        "links": {
          "$ref": "#/definitions/ResponseStatusLinks"
        }
      }
    },
    "ResponseStatusLinks": {
      "required": [
        "status"
      ],
      "properties": {
        "status": {
          "$ref": "#/definitions/Href"
        }
      }
    },
...
    "Href": {
      "required": [
        "href"
      ],
      "properties": {
        "href": {
          "type": "string"
        }
      }
    },
...
    "GetAsset": {
      "required": [
        "id",
        "created",
        "created_by",
        "last_modified",
        "last_modified_by",
        "links"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "external_ids": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "created": {
          "type": "string",
          "format": "date-time"
        },
        "created_by": {
          "type": "string"
        },
        "last_modified": {
          "type": "string",
          "format": "date-time"
        },
        "last_modified_by": {
          "type": "string"
        },
        "links": {
          "$ref": "#/definitions/AssetLinks"
        }
      }
    },
    "AssetLinks": {
      "required": [
        "self",
        "related",
        "parent",
        "children",
        "variants",
        "metadata"
      ],
      "properties": {
        "self": {
          "$ref": "#/definitions/Href"
        },
        "related": {
          "$ref": "#/definitions/Href"
        },
        "parent": {
          "$ref": "#/definitions/Href"
        },
        "children": {
          "$ref": "#/definitions/Href"
        },
        "variants": {
          "$ref": "#/definitions/Href"
        },
        "metadata": {
          "$ref": "#/definitions/Href"
        }
      }
    },
...
  }
}